refactor(motion): 重构运动控制模块
- 更新 Encoders 和 Motors 类的构造函数，使用固定大小的引脚数组
- 移除不必要的参数检查
- 修改电机速度类型为 int16_t
- 重构示例代码，使用新的 Motion 模块
- 优化编码器和电机的初始化及使用

+++ lib/Motor/Motor.hpp
@@ -29,7 +29,7 @@ class Motor {
          * @brief 设置电机的转速
          * @param rate 
          */
-        void setSpeed(int8_t rate) {
+        void setSpeed(int16_t rate) {
             int8_t pwmValue = map(abs(rate), 0, 100, _THR_MIN, _THR_MAX);
             pwmValue = constrain(pwmValue, _THR_MIN, _THR_MAX);
     
@@ -47,7 +47,7 @@ class Motor {
 
         /**
          * @brief 设置电机的上下限
-         * @param rate 
+         * @param THR_MIN, THR_MAX 
          */
         void setSpeedLimit(uint16_t THR_MIN, uint16_t THR_MAX) {
             _THR_MIN = THR_MIN;
